Jam integration various warning fixes

+++ db/log_writer.cc
@@ -104,7 +104,7 @@ Status Writer::EmitPhysicalRecord(RecordType t, const char* ptr, size_t n) {
       s = dest_->Flush();
     }
   }
-  block_offset_ += kHeaderSize + n;
+  block_offset_ += kHeaderSize + (int)n;
   return s;
 }
